basic-platforming - added ground, no collision.
Continuing the trend of updating the game in baby steps. Another
FlxSprite-inheriting class is now instantiated in the PlayState, ripe
for the player to collide with.

Also did some minor refactor. It doesn't actually make sense to have an
entities FlxGroup, given that it'll just hold everything made for the
PlayState (which is what the Playstate is supposed to contain, anyway).
There may be reasons to group things differently in the future, but I
think this makes a fine amount of sense for now.