basic-platforming - added platform collision.
Also touched up some of the update logic in the process. If you're
looking at this commit for reference for some other project, I recommend
digging deeper into what the FlxG.collide() method is actually doing.
It's one of the larger black-box type methods we directly invoke, and
having a thorough understanding of it will pay off greatly later.