auto-run - implemented auto player movement.
The current implementation isn't all that clean code-wise. We're
layering on complexity within the Player class in a way that's going to
make future changes progressively more difficult and risky. That will be
fixed in a future branch, for now keeping the code more straightforward
to read down the page is of more value for students, I think.

Responsiveness of movement control feels pretty good to me, I'll plan on
messing around with jump control in an upcoming commit.