auto-run - added basic walljump.
Implemented a really basic walljump for the player. Code isn't very well
architected, but it's functional enough.

I was very easily amused by this stage of the prototype. Just trying out
different walljumps and experiencing surprise when I found a way to do a
new one. It is possible to jump off of every wall in this version of the
prototype (even the off-screen ones!)

If you'd like to challenge yourself, see if you can find methods to do
each unique walljump. Maybe we should race to see who can do it the
fastest.